extracted github related functionality to its own submodule
